requirements.txt
Celery
Django
gunicorn
psycopg[binary]
dj-database-url
django-celery-beat 
django-celery-results
python-decouple
redis
jupyter
selenium
beautifulsoup4
openai
djangorestframework
django-cors-headers